Added stuff learned in class

+++ src/Move.java
@@ -2,7 +2,7 @@ import java.awt.*;
 
 public interface Move {
     void paint(Graphics g);
-    void setPosition(double x, double y);
+    void setPosition(int x, int y);
     void move();
     Rectangle getBounds();
 }
